fix(runtime): keep the demo rides' FIT totals at sub-second precision
The totals were cast to a Duration before being scaled by time.Second, so
they truncated to whole seconds and disagreed with the record timestamps
beside them, which the same file already scales the other way round.

Also names the rides in Seed's failure message, and settles what CI's lint
caught: seeding the rides reused the outer err and so extended its live
range past three inner declarations that had shadowed it harmlessly until
now, plus an unlambda and a field ordering in the new tests.

+++ internal/demo/activities.go
@@ -166,8 +166,8 @@ func (s *rideSpec) ride(stages []route.Route, now time.Time) (Ride, error) {
 		FIT: activity.FIT{
 			RecordingDevice:     "Demo Head Unit",
 			Records:             records,
-			TotalTimerTime:      time.Duration(summary.MovingSeconds) * time.Second,
-			TotalElapsedTime:    time.Duration(summary.ElapsedSeconds) * time.Second,
+			TotalTimerTime:      time.Duration(summary.MovingSeconds * float64(time.Second)),
+			TotalElapsedTime:    time.Duration(summary.ElapsedSeconds * float64(time.Second)),
 			TotalAscentMetres:   summary.AscentMetres,
 			HasTotalTimerTime:   true,
 			HasTotalElapsedTime: true,
